Moves some data classes into `WarClan`
